feat: oai 6.8.1
