+++ package.json
@@ -45,7 +45,7 @@
   "dependencies": {
     "agentkeepalive": "^4.6.0",
     "dotenv": "^16.5.0",
-    "openai": "6.3.0",
+    "openai": "6.8.1",
     "ws": "^8.18.2"
   }
 }
